split question service from user and organise routes
