+++ backend/QuestionService copy/.env.example
@@ -1,0 +1,5 @@
+MONGODB_URI="put_your_own_mongodb_uri_here"
+SECRET_KEY="generate_your_own_secret_key"
+PORT="8080"
+BACKEND_URL="http://localhost:8080"
+FRONTEND_URL="http://localhost:5173"
