+++ backend/QuestionService copy/.gitignore
@@ -1,0 +1,3 @@
+.env
+backend.exe
+backend.exe~
