+++ backend/QuestionService copy/database/databaseConnection.go
@@ -1,0 +1,48 @@
+package database
+
+import (
+	"context"
+	"fmt"
+	"log"
+	"os"
+	"time"
+
+	"github.com/joho/godotenv"
+	"go.mongodb.org/mongo-driver/mongo"
+	"go.mongodb.org/mongo-driver/mongo/options"
+)
+
+// DBinstance func
+func DBinstance() *mongo.Client {
+	err := godotenv.Load(".env") // Load .env file
+
+	if err != nil {
+		log.Fatal("Error loading .env file")
+	}
+
+	uri := os.Getenv("MONGODB_URI")
+
+	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second) // timeout if does not connect within 10 seconds
+
+	client, err := mongo.Connect(ctx, options.Client().
+		ApplyURI(uri))
+
+	defer cancel()
+	if err != nil {
+		log.Fatal(err)
+	}
+	fmt.Println("Connected to MongoDB!")
+
+	return client
+}
+
+// Client Database instance
+var Client *mongo.Client = DBinstance()
+
+// OpenCollection is a  function makes a connection with a collection in the database
+func OpenCollection(client *mongo.Client, databaseName string, collectionName string) *mongo.Collection {
+
+	var collection *mongo.Collection = client.Database(databaseName).Collection(collectionName)
+
+	return collection
+}
