+++ backend/QuestionService copy/main.go
@@ -1,0 +1,50 @@
+package main
+
+import (
+	"os"
+
+	"backend/middleware"
+	routes "backend/routes"
+
+	"github.com/gin-contrib/cors"
+	"github.com/gin-gonic/gin"
+)
+
+func main() {
+	port := os.Getenv("PORT")
+
+	if port == "" {
+		port = "8000"
+	}
+
+	router := gin.New()
+	router.Use(gin.Logger())
+
+	// Apply CORS middleware with custom configuration
+	router.Use(cors.New(cors.Config{
+		AllowOrigins:     []string{"http://localhost:5173"}, // Ensure it matches your frontend port
+		AllowMethods:     []string{"POST", "GET", "OPTIONS", "PUT", "DELETE"},
+		AllowHeaders:     []string{"Content-Type", "Authorization", "token"},
+		ExposeHeaders:    []string{"Content-Length"},
+		AllowCredentials: true,
+		MaxAge:           12 * 60 * 60,
+	}))
+
+	routes.UserRoutes(router) // Creates User api routes
+
+	router.Use(middleware.Authentication())
+
+	routes.QuestionRoutes(router) // Creates Question api routes
+	routes.ProfileRoutes(router)  // Creates Profile api routes
+	router.GET("/api-1", func(c *gin.Context) {
+		c.JSON(200, gin.H{"success": "Access granted for api-1"})
+	})
+
+	// API-2
+	router.GET("/api-2", func(c *gin.Context) {
+		c.JSON(200, gin.H{"success": "Access granted for api-2"})
+	})
+
+	router.Run(":" + port)
+
+}
