+++ backend/QuestionService/middleware/authMiddleware.go
@@ -1,45 +1,70 @@
 package middleware
 
 import (
-	"fmt"
+	"encoding/json"
+	"io"
 	"net/http"
-
-	helper "backend/helpers"
+	"os"
+	"strings"
 
 	"github.com/gin-gonic/gin"
 )
 
+
+type Response struct {
+	Message string `json:"message"`
+}
+
 // Authz validates token and authorizes users
 func Authentication() gin.HandlerFunc {
 	return func(c *gin.Context) {
-		fmt.Println("Authenticating")
-		var clientToken string = c.Request.Header.Get("token")
+		AUTH_URL := os.Getenv("AUTH_URL");
 
-		if clientToken == "" {
-			c.JSON(http.StatusInternalServerError, gin.H{"error": fmt.Sprintf("No Authorization header provided")})
+		clientToken := c.Request.Header.Get("Authorization")
+		strArr := strings.Split(clientToken, "Bearer ")
+
+		if clientToken == "" || len(strArr) < 2 {
+			c.JSON(http.StatusUnauthorized, gin.H{"error": gin.H{ "message": "No token found" }})
 			c.Abort()
 			return
 		}
 
-		claims, err := helper.ValidateToken(clientToken)
+		bearer := "Bearer " + strArr[1]
 
-		if err != "" {
-			c.JSON(http.StatusInternalServerError, gin.H{"error": err})
+		req, err := http.NewRequest("GET", AUTH_URL, nil)
+		if err != nil {
+			c.JSON(http.StatusInternalServerError, gin.H{"error": gin.H{ "message": err.Error() }})
 			c.Abort()
 			return
 		}
 
-		if claims == nil {
-			c.JSON(http.StatusInternalServerError, gin.H{"error": err})
+		req.Header.Add("Authorization", bearer)
+
+		client := &http.Client{}
+		resp, err := client.Do(req)
+
+		if err != nil {
+			c.JSON(http.StatusInternalServerError, gin.H{"error": gin.H{ "message": err.Error() }})
 			c.Abort()
 			return
 		}
 
-		// Stores the user information in the context
-		// c.Set("email", claims.Email)
-		// c.Set("username", claims.Username)
-		c.Set("uid", claims.Uid)
+		defer resp.Body.Close()
+
+		if resp.StatusCode != http.StatusOK {	
+			respBytes, _ := io.ReadAll(resp.Body)
+			var response Response
+			err := json.Unmarshal(respBytes, &response)
+			if err != nil {
+				c.JSON(http.StatusUnauthorized, gin.H{"error": gin.H{ "message": err.Error() }})
+				c.Abort()
+				return
+			}
+
+			c.JSON(http.StatusInternalServerError, gin.H{"error": gin.H{ "message": response.Message }})	
+			c.Abort()
+		}
 
 		c.Next()
-	}
+	}	
 }
